+++ plugins/woocommerce-admin/client/analytics/components/leaderboard/test/index.js
@@ -6,33 +6,41 @@ import { mount, shallow } from 'enzyme';
 /**
  * WooCommerce dependencies
  */
-import { formatCurrency, getCurrencyFormatDecimal } from 'lib/currency-format';
-import { numberFormat } from 'lib/number-format';
+import { numberFormat } from '@woocommerce/number';
+import Currency from '@woocommerce/currency';
 
 /**
  * Internal dependencies
  */
 import { Leaderboard } from '../';
 import mockData from '../data/top-selling-products-mock-data';
+import { CURRENCY } from '@woocommerce/wc-admin-settings';
+
+const { formatCurrency, formatDecimal } = Currency( CURRENCY );
 
 const rows = mockData.map( ( row ) => {
-	const { name, items_sold: itemsSold, net_revenue: netRevenue, orders_count: ordersCount } = row;
+	const {
+		name,
+		items_sold: itemsSold,
+		net_revenue: netRevenue,
+		orders_count: ordersCount,
+	} = row;
 	return [
 		{
 			display: '<a href="#">' + name + '</a>',
 			value: name,
 		},
 		{
-			display: numberFormat( itemsSold ),
+			display: numberFormat( CURRENCY, itemsSold ),
 			value: itemsSold,
 		},
 		{
-			display: numberFormat( ordersCount ),
+			display: numberFormat( CURRENCY, ordersCount ),
 			value: ordersCount,
 		},
 		{
 			display: formatCurrency( netRevenue ),
-			value: getCurrencyFormatDecimal( netRevenue ),
+			value: formatDecimal( netRevenue ),
 		},
 	];
 } );
@@ -85,7 +93,7 @@ describe( 'Leaderboard', () => {
 		expect( firstRow[ 1 ].value ).toBe( mockData[ 0 ].items_sold );
 		expect( firstRow[ 2 ].value ).toBe( mockData[ 0 ].orders_count );
 		expect( firstRow[ 3 ].value ).toBe(
-			getCurrencyFormatDecimal( mockData[ 0 ].net_revenue )
+			formatDecimal( mockData[ 0 ].net_revenue )
 		);
 
 		expect(
@@ -93,10 +101,10 @@ describe( 'Leaderboard', () => {
 		).toBe( 5 );
 		expect( tableItems.at( 0 ).text() ).toBe( mockData[ 0 ].name );
 		expect( tableItems.at( 1 ).text() ).toBe(
-			numberFormat( mockData[ 0 ].items_sold )
+			numberFormat( CURRENCY, mockData[ 0 ].items_sold )
 		);
 		expect( tableItems.at( 2 ).text() ).toBe(
-			numberFormat( mockData[ 0 ].orders_count )
+			numberFormat( CURRENCY, mockData[ 0 ].orders_count )
 		);
 		expect( tableItems.at( 3 ).text() ).toBe(
 			formatCurrency( mockData[ 0 ].net_revenue )