+++ plugins/woocommerce-admin/client/analytics/report/coupons/table.js
@@ -10,17 +10,17 @@ import { map } from 'lodash';
  */
 import { Date, Link } from '@woocommerce/components';
 import { defaultTableDateFormat } from 'lib/date';
-import { formatCurrency, getCurrencyFormatDecimal } from 'lib/currency-format';
 import { getNewPath, getPersistedQuery } from '@woocommerce/navigation';
-import { formatValue } from 'lib/number-format';
+import { formatValue } from '@woocommerce/number';
 import { getSetting } from '@woocommerce/wc-admin-settings';
 
 /**
  * Internal dependencies
  */
 import ReportTable from 'analytics/components/report-table';
+import { CurrencyContext } from 'lib/currency-context';
 
-export default class CouponsReportTable extends Component {
+class CouponsReportTable extends Component {
 	constructor() {
 		super();
 
@@ -71,9 +71,18 @@ export default class CouponsReportTable extends Component {
 		const { query } = this.props;
 		const persistedQuery = getPersistedQuery( query );
 		const dateFormat = getSetting( 'dateFormat', defaultTableDateFormat );
+		const {
+			formatCurrency,
+			formatDecimal: getCurrencyFormatDecimal,
+			getCurrency,
+		} = this.context;
 
 		return map( coupons, ( coupon ) => {
-			const { amount, coupon_id: couponId, orders_count: ordersCount } = coupon;
+			const {
+				amount,
+				coupon_id: couponId,
+				orders_count: ordersCount,
+			} = coupon;
 			const extendedInfo = coupon.extended_info || {};
 			const {
 				code,
@@ -102,7 +111,7 @@ export default class CouponsReportTable extends Component {
 			} );
 			const ordersLink = (
 				<Link href={ ordersUrl } type="wc-admin">
-					{ formatValue( 'number', ordersCount ) }
+					{ formatValue( getCurrency(), 'number', ordersCount ) }
 				</Link>
 			);
 
@@ -148,7 +157,13 @@ export default class CouponsReportTable extends Component {
 	}
 
 	getSummary( totals ) {
-		const { coupons_count: couponsCount = 0, orders_count: ordersCount = 0, amount = 0 } = totals;
+		const {
+			coupons_count: couponsCount = 0,
+			orders_count: ordersCount = 0,
+			amount = 0,
+		} = totals;
+		const { formatCurrency, getCurrency } = this.context;
+		const currency = getCurrency();
 		return [
 			{
 				label: _n(
@@ -157,7 +172,7 @@ export default class CouponsReportTable extends Component {
 					couponsCount,
 					'woocommerce-admin'
 				),
-				value: formatValue( 'number', couponsCount ),
+				value: formatValue( currency, 'number', couponsCount ),
 			},
 			{
 				label: _n(
@@ -166,7 +181,7 @@ export default class CouponsReportTable extends Component {
 					ordersCount,
 					'woocommerce-admin'
 				),
-				value: formatValue( 'number', ordersCount ),
+				value: formatValue( currency, 'number', ordersCount ),
 			},
 			{
 				label: __( 'amount discounted', 'woocommerce-admin' ),
@@ -216,3 +231,7 @@ export default class CouponsReportTable extends Component {
 		);
 	}
 }
+
+CouponsReportTable.contextType = CurrencyContext;
+
+export default CouponsReportTable;