+++ plugins/woocommerce-admin/client/analytics/report/index.js
@@ -31,6 +31,10 @@ import CustomersReport from './customers';
 import ReportError from 'analytics/components/report-error';
 import { searchItemsByString } from 'wc-api/items/utils';
 import withSelect from 'wc-api/with-select';
+import {
+	CurrencyContext,
+	getFilteredCurrencyInstance,
+} from 'lib/currency-context';
 
 export const REPORTS_FILTER = 'woocommerce_admin_reports_list';
 const manageStock = getSetting( 'manageStock', 'no' );
@@ -142,7 +146,13 @@ class Report extends Component {
 			return null;
 		}
 		const Container = report.component;
-		return <Container { ...this.props } />;
+		return (
+			<CurrencyContext.Provider
+				value={ getFilteredCurrencyInstance( getQuery() ) }
+			>
+				<Container { ...this.props } />
+			</CurrencyContext.Provider>
+		);
 	}
 }