+++ plugins/woocommerce-admin/client/analytics/report/products/table.js
@@ -10,14 +10,9 @@ import { map } from 'lodash';
 /**
  * WooCommerce dependencies
  */
-import {
-	formatCurrency,
-	getCurrencyFormatDecimal,
-	renderCurrency,
-} from 'lib/currency-format';
 import { getNewPath, getPersistedQuery } from '@woocommerce/navigation';
 import { Link, Tag } from '@woocommerce/components';
-import { formatValue } from 'lib/number-format';
+import { formatValue } from '@woocommerce/number';
 import { getAdminLink, getSetting } from '@woocommerce/wc-admin-settings';
 
 /**
@@ -27,6 +22,7 @@ import CategoryBreacrumbs from '../categories/breadcrumbs';
 import { isLowStock } from './utils';
 import ReportTable from 'analytics/components/report-table';
 import withSelect from 'wc-api/with-select';
+import { CurrencyContext } from 'lib/currency-context';
 import './style.scss';
 
 const manageStock = getSetting( 'manageStock', 'no' );
@@ -38,6 +34,7 @@ class ProductsReportTable extends Component {
 
 		this.getHeadersContent = this.getHeadersContent.bind( this );
 		this.getRowsContent = this.getRowsContent.bind( this );
+		this.getSummary = this.getSummary.bind( this );
 	}
 
 	getHeadersContent() {
@@ -105,6 +102,12 @@ class ProductsReportTable extends Component {
 	getRowsContent( data = [] ) {
 		const { query } = this.props;
 		const persistedQuery = getPersistedQuery( query );
+		const {
+			render: renderCurrency,
+			formatDecimal: getCurrencyFormatDecimal,
+			getCurrency,
+		} = this.context;
+		const currency = getCurrency();
 
 		return map( data, ( row ) => {
 			const {
@@ -185,7 +188,7 @@ class ProductsReportTable extends Component {
 					value: sku,
 				},
 				{
-					display: formatValue( 'number', itemsSold ),
+					display: formatValue( currency, 'number', itemsSold ),
 					value: itemsSold,
 				},
 				{
@@ -238,7 +241,11 @@ class ProductsReportTable extends Component {
 						.join( ', ' ),
 				},
 				{
-					display: formatValue( 'number', variations.length ),
+					display: formatValue(
+						currency,
+						'number',
+						variations.length
+					),
 					value: variations.length,
 				},
 				manageStock === 'yes'
@@ -254,7 +261,11 @@ class ProductsReportTable extends Component {
 				manageStock === 'yes'
 					? {
 							display: extendedInfoManageStock
-								? formatValue( 'number', stockQuantity )
+								? formatValue(
+										currency,
+										'number',
+										stockQuantity
+								  )
 								: __( 'N/A', 'woocommerce-admin' ),
 							value: stockQuantity,
 					  }
@@ -270,6 +281,8 @@ class ProductsReportTable extends Component {
 			net_revenue: netRevenue = 0,
 			orders_count: ordersCount = 0,
 		} = totals;
+		const { formatCurrency, getCurrency } = this.context;
+		const currency = getCurrency();
 		return [
 			{
 				label: _n(
@@ -278,7 +291,7 @@ class ProductsReportTable extends Component {
 					productsCount,
 					'woocommerce-admin'
 				),
-				value: formatValue( 'number', productsCount ),
+				value: formatValue( currency, 'number', productsCount ),
 			},
 			{
 				label: _n(
@@ -287,7 +300,7 @@ class ProductsReportTable extends Component {
 					itemsSold,
 					'woocommerce-admin'
 				),
-				value: formatValue( 'number', itemsSold ),
+				value: formatValue( currency, 'number', itemsSold ),
 			},
 			{
 				label: __( 'net sales', 'woocommerce-admin' ),
@@ -300,7 +313,7 @@ class ProductsReportTable extends Component {
 					ordersCount,
 					'woocommerce-admin'
 				),
-				value: formatValue( 'number', ordersCount ),
+				value: formatValue( currency, 'number', ordersCount ),
 			},
 		];
 	}
@@ -359,6 +372,8 @@ class ProductsReportTable extends Component {
 		);
 	}
 }
+
+ProductsReportTable.contextType = CurrencyContext;
 
 export default compose(
 	withSelect( ( select, props ) => {